[Title] [MadrasaApp] Refactored LoginScreen composable structure [Cause] LoginScreen composable had tightly coupled UI elements and could be better structured for readability and maintainability. [Countermeasure] - Extracted smaller composables (e.g., RoleSelection, etc.) - Applied state hoisting for input and dropdown state - Improved naming consistency - Cleaned up navigation callback handling for LoginScreen
